pyproject.toml
[project]
name = "sheetsearchai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "google-api-python-client>=2.190.0",
    "openai>=2.21.0",
    "openpyxl>=3.1.5",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
]
